# apps/api/openapi.json
{
  "components": {
...
    "securitySchemes": { "bearerAuth": { "scheme": "bearer", "type": "http" } }
  },
  "info": { "title": "FF RESTaurent API", "version": "1.1.0" },
  "openapi": "3.0.3",
  "paths": {
...
    "/participant-groups": {
      "get": {
        "operationId": "getParticipant-groups",
        "responses": {
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "operationId": "postParticipant-groups",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "memberIds": {
                    "items": { "minLength": 1, "type": "string" },
                    "maxItems": 100,
                    "minItems": 2,
                    "type": "array"
                  },
                  "name": { "maxLength": 80, "minLength": 1, "type": "string" }
                },
                "required": ["name", "memberIds"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/participant-groups/{id}": {
      "delete": {
        "operationId": "deleteParticipant-groupsById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "minLength": 1, "type": "string" }
          }
        ],
        "responses": {
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "operationId": "putParticipant-groupsById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "minLength": 1, "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "memberIds": {
                    "items": { "minLength": 1, "type": "string" },
                    "maxItems": 100,
                    "minItems": 2,
                    "type": "array"
                  },
                  "name": { "maxLength": 80, "minLength": 1, "type": "string" }
                },
                "required": ["name", "memberIds"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "2XX": {
            "content": { "application/json": { "schema": {} } },
            "description": "Default Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "code": { "type": "string" },
                    "message": { "type": "string" }
                  },
                  "required": ["message"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  }
}